% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R, R/mean_function.R
\name{calculate_mean}
\alias{calculate_mean}
\title{Calculate the Mean of a Numeric Vector}
\usage{
calculate_mean(x)

calculate_mean(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
The mean of the vector

The mean of the vector
}
\description{
Calculate the Mean of a Numeric Vector

Calculate the Mean of a Numeric Vector
}
