% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_function.R
\name{calculate_median}
\alias{calculate_median}
\title{Calculate the Median of a Numeric Vector}
\usage{
calculate_median(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
The median of the vector
}
\description{
Calculate the Median of a Numeric Vector
}
